basic detox support

+++ util/test-e2e.sh
@@ -1,18 +1,28 @@
 #!/bin/bash
 
-$rnxRoot/util/killProcess.sh appium
-
-if [ "${IS_BUILD_AGENT}" == true ] || [ "${1}" == "release" ]; then  
-  $rnxRoot/util/killProcess.sh fake-server # kill other fake servers in the CI to clear port 3000
-  nohup npm run fake-server &
-else
-  # disconnect hardware keyboard from simulator
-  osascript $rnxRoot/util/set_hardware_keyboard.applescript 0
-fi
-
 echo "##teamcity[blockOpened name='E2E']"
 set +e
-BABEL_ENV=specs specFilterString=./test/e2e/*.e2e.spec.js node ./test/e2e/support/jasmine-runner.js
+
+#todo deprecate jasmine/appium support eventually
+if [ -d "./test/e2e/support/jasmine-runner.js" ]; then
+  $rnxRoot/util/killProcess.sh appium
+
+  if [ "${IS_BUILD_AGENT}" == true ] || [ "${1}" == "release" ]; then  
+    $rnxRoot/util/killProcess.sh fake-server # kill other fake servers in the CI to clear port 3000
+    nohup npm run fake-server &
+  else
+    # disconnect hardware keyboard from simulator
+    osascript $rnxRoot/util/set_hardware_keyboard.applescript 0
+  fi
+
+  BABEL_ENV=specs specFilterString=./test/e2e/*.e2e.spec.js node ./test/e2e/support/jasmine-runner.js
+
+else 
+  $rnxRoot/util/killProcess.sh detox-server
+  ./node_modules/.bin/detox-server &
+  BABEL_ENV=specs mocha test/e2e --opts ./test/e2e/mocha.opts
+fi
+
 exitCode=$?
 set -e
 echo "##teamcity[blockClosed name='E2E']"
